requirements.txt
# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Numerical Computing  
numpy>=1.26.0
pandas>=2.0.0

# Deep Learning
tensorflow>=2.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Dashboard
streamlit>=1.28.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
